+++ src/components/Measurement.scss
@@ -28,3 +28,4 @@
   color: $gray;
   font-size: 0.8em;
 }
+
