+++ src/components/graphs/PastGraph.scss
@@ -1,0 +1,45 @@
+/*body { font: 13px sans-serif; }*/
+
+rect { fill: #fff; }
+
+/*ul {
+  list-style-type: none;
+  margin: 0.5em 0em 0.5em 0em;
+  width: 100%; }
+
+ul li {
+  display: table-cell;
+  vertical-align: middle;
+  margin: 0em;
+  padding: 0em 1em; }*/
+
+.axis { font-size: 1.5em; }
+
+.chart {
+  /*background-color: #F7F2C5;*/
+  width: 960px;
+  height: 500px; }
+
+circle, .line {
+  fill: none;
+  stroke: steelblue;
+  stroke-width: 2px; }
+
+circle {
+  fill: white;
+  fill-opacity: 0.2;
+  cursor: move; }
+
+circle.selected {
+  fill: #ff7f0e;
+  stroke: #ff7f0e; }
+
+circle:hover {
+  fill: #ff7f0e;
+  stroke: #707f0e; }
+
+circle.selected:hover {
+  fill: #ff7f0e;
+  stroke: #ff7f0e; }
+
+
